modularizing clustering script
